# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "AssignmentEnd": {
        "additionalProperties": false,
        "description": "Close an assignment on a date. The row stays; only its end moves.",
        "properties": {
          "effective_to": {
            "format": "date",
            "title": "Effective To",
            "type": "string"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          }
        },
        "required": [
          "effective_to",
          "expected_version"
        ],
        "title": "AssignmentEnd",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "properties": {
          "code": {
            "description": "Stable machine-readable code, safe to branch on.",
            "title": "Code",
            "type": "string"
          },
          "correlation_id": {
            "default": "",
            "title": "Correlation Id",
            "type": "string"
          },
          "field_errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "title": "Field Errors",
            "type": "array"
          },
          "message": {
            "description": "Safe to display to a person.",
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "TreeRead": {
        "description": "The whole tree, flat.\n\nFlat rather than nested, deliberately. The client builds the nesting from `parent_id`, which\nmeans one shape serves the tree view, a search result and a partially expanded view without\nthe server guessing which is wanted. A nested response would also have to pick a depth limit,\nand any limit is wrong for somebody.",
        "properties": {
          "as_at": {
            "format": "date",
            "title": "As At",
            "type": "string"
          },
          "is_empty": {
            "title": "Is Empty",
            "type": "boolean"
          },
          "units": {
            "items": {
              "$ref": "#/components/schemas/OrgUnitRead"
            },
            "title": "Units",
            "type": "array"
          }
        },
        "required": [
          "as_at",
          "units",
          "is_empty"
        ],
        "title": "TreeRead",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "Governed human and AI work operating system",
    "title": "UBOSS AI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/hierarchy": {
      "get": {
        "description": "Units and positions, flat, as at one date.\n\nFlat rather than nested: the client builds the nesting from `parent_id`, which means one\nresponse shape serves the tree, a search result and a partially expanded view. A nested\nresponse would also have to pick a depth limit, and every limit is wrong for somebody.\n\n`is_empty` distinguishes \"this organisation has no tree yet\" from \"the request failed\" \u2014 two\nstates that look identical on screen and mean opposite things.",
        "operationId": "read_tree_api_v1_hierarchy_get",
        "parameters": [
          {
            "description": "Show the structure as it stands on this date. Defaults to today.",
            "in": "query",
            "name": "as_at",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Show the structure as it stands on this date. Defaults to today.",
              "title": "As At"
            }
          },
          {
            "in": "query",
            "name": "include_archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Archived",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TreeRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not signed in, or the session ended."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Refused. The reason is in the audit trail."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No such record, for this caller."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The record moved, or the key was reused."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Some field was not accepted."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too many attempts. See Retry-After."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A fault on our side. Nothing was changed."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A dependency did not answer. Retryable."
          }
        },
        "summary": "The company tree",
        "tags": [
          "hierarchy"
        ]
      }
    },
    "/api/v1/hierarchy/assignments/{assignment_id}": {
      "patch": {
        "description": "Closes it on a date. Never a delete \u2014 they held the seat, and that stays true.",
        "operationId": "end_assignment_api_v1_hierarchy_assignments__assignment_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "assignment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Assignment Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignmentEnd"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response End Assignment Api V1 Hierarchy Assignments  Assignment Id  Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not signed in, or the session ended."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Refused. The reason is in the audit trail."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No such record, for this caller."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The record moved, or the key was reused."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Some field was not accepted."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too many attempts. See Retry-After."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A fault on our side. Nothing was changed."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A dependency did not answer. Retryable."
          }
        },
        "summary": "End an assignment",
        "tags": [
          "hierarchy"
        ]
      }
    },
...
    "/api/v1/hierarchy/issues": {
      "get": {
        "description": "Orphaned managers and vacant seats.\n\nNot cycles or duplicate identifiers: those are refused by the database and can never be here\nto find. These are states a real restructure passes through, so they are reported rather than\nrefused, and a person decides what to do about them.",
        "operationId": "read_issues_api_v1_hierarchy_issues_get",
        "parameters": [
          {
            "description": "Show the structure as it stands on this date. Defaults to today.",
            "in": "query",
            "name": "as_at",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Show the structure as it stands on this date. Defaults to today.",
              "title": "As At"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ValidationIssue"
                  },
                  "title": "Response Read Issues Api V1 Hierarchy Issues Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not signed in, or the session ended."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Refused. The reason is in the audit trail."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No such record, for this caller."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The record moved, or the key was reused."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Some field was not accepted."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too many attempts. See Retry-After."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A fault on our side. Nothing was changed."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A dependency did not answer. Retryable."
          }
        },
        "summary": "Problems worth someone's attention",
        "tags": [
          "hierarchy"
        ]
      }
    },
    "/api/v1/hierarchy/positions": {
      "post": {
        "description": "Create a seat. It exists whether or not anybody is in it \u2014 PLAN \u00a75's vacant positions.",
        "operationId": "create_position_api_v1_hierarchy_positions_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Create Position Api V1 Hierarchy Positions Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not signed in, or the session ended."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Refused. The reason is in the audit trail."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No such record, for this caller."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The record moved, or the key was reused."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Some field was not accepted."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too many attempts. See Retry-After."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A fault on our side. Nothing was changed."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A dependency did not answer. Retryable."
          }
        },
        "summary": "Add a position",
        "tags": [
          "hierarchy"
        ]
      }
    },
...
  }
}